requirements.txt
pandas
streamlit
plotly
python-dotenv
psycopg2-binary
requests
